Execute one of a set of bayesian models

Execute one of a set of bayesian models

Usage

bayesian_true_prevalence_model(..., model_type = c("logit", "simpler"))

Arguments

...

Arguments passed on to bayesian_component_simpler_model, bayesian_component_logit_model

cache_result

save the result of the sampling in memory for the current session

pos_obs

the number of positive observations for a given test

n_obs

the number of observations for a given test

false_pos_controls

the number of positives that appeared in the specificity disease-free control group. These are by definition false positives. This is (1-specificity)*n_controls

n_controls

the number of controls in the specificity disease-free control group.

false_neg_diseased

the number of negatives that appeared in the sensitivity confirmed disease group. These are by definition false negatives. This is (1-sensitivity)*n_diseased

n_diseased

the number of confirmed disease cases in the sensitivity control group.

sens

the prior sensitivity of the test as a beta_dist.

spec

the prior specificity of the test as a beta_dist.

confint

confidence interval limits

fmt

a sprintf formatting string accepting 3 numbers

chains

A positive integer specifying the number of Markov chains. The default is 4.

iter

A positive integer specifying the number of iterations for each chain (including warmup). The default is 2000.

warmup

A positive integer specifying the number of warmup (aka burnin) iterations per chain. If step-size adaptation is on (which it is by default), this also controls the number of iterations for which adaptation is run (and hence these warmup samples should not be used for inference). The number of warmup iterations should be smaller than iter and the default is iter/2.

model_type

The bayesian model used - one of "logit" or "simpler"

Value

A dataframe containing the following columns:

test (character) - the name of the test or panel

prevalence.lower (numeric) - the lower estimate

prevalence.median (numeric) - the median estimate

prevalence.upper (numeric) - the upper estimate

prevalence.method (character) - the method of estimation

prevalence.label (character) - a fomatted label of the true prevalence estimate with CI

Ungrouped.

No default value.