Significance of an uncertain test result

Calculates a p-value for a count of positive test results based on false positive (specificity) controls. The null hypothesis is that the prevalence of the disease is zero.

Usage

fp_p_value(
  pos_obs,
  n_obs,
  false_pos_controls,
  n_controls,
  format = "%1.3g",
  lim = 1e-04,
  bonferroni = NULL,
  ...
)

Arguments

pos_obs

the number of positive observations for a given test

n_obs

the number of observations for a given test

false_pos_controls

the number of positives that appeared in the specificity disease-free control group. These are by definition false positives. This is (1-specificity)*n_controls

n_controls

the number of controls in the specificity disease-free control group.

format

a sprintf fmt string for the p-value

lim

a lower value to display

bonferroni

the number of simultaneous hypotheses that are being tested

...

not used

Value

a vector of p-values for the count

Details

This p_value does not tell you whether this count can be trusted only if the prevalence of this disease is significantly more than zero after this observation.

Examples

# calculate p-values for counts derived from 300 samples
# 10 observations is within noise of test
# 20 observations is unlikely on 1200 observations
fp_p_value(c(10,2,4,3,10,20), 1200, c(0,0,2,0,2,0)+2, 800)
#> [1] "0.0571"     "0.682"      "0.733"      "0.544"      "0.216"     
#> [6] "0.000956 †"

# if the same observations are made against a smaller group then we get 
# a positive result for 10
fp_p_value( c(10,2,4,3,10,20), 1000, c(2,2,4,2,4,2), 800)
#> [1] "0.0289"     "0.603"      "0.634"      "0.452"      "0.128"     
#> [6] "0.000224 †"

tibble::tibble(
  x = c(1,2,5,10,20,40,20,20,20,20,20),
  n = 1000,
  fp_controls = c(0,0,0,0,0,0,0,1,2,3,4)+2,
  n_controls = 800
) %>% dplyr::mutate(
  p_value = fp_p_value(x, n, fp_controls, n_controls)
) %>% dplyr::glimpse()
#> Rows: 11
#> Columns: 5
#> $ x           <dbl> 1, 2, 5, 10, 20, 40, 20, 20, 20, 20, 20
#> $ n           <dbl> 1000, 1000, 1000, 1000, 1000, 1000, 1000, 1000, 1000, 1000…
#> $ fp_controls <dbl> 2, 2, 2, 2, 2, 2, 2, 3, 4, 5, 6
#> $ n_controls  <dbl> 800, 800, 800, 800, 800, 800, 800, 800, 800, 800, 800
#> $ p_value     <chr> "0.766", "0.603", "0.229", "0.0289", "0.000224 †", "<0.000…