Calculate a vaccine effectiveness estimate based on an odds ratio

This assumes that OR ~ RR which is only true if controls >> cases The OR method can be used in test negative designs where disease positive relates to vaccine treatable disease and disease negative relates to non vaccine treatable disease

This assumes that OR ~ RR which is only true if controls >> cases The OR method can be used in test negative designs where disease positive relates to vaccine treatable disease and disease negative relates to non vaccine treatable disease

Usage

odds_ratio_ve(
  vaccinatedCase,
  unvaccinatedCase,
  vaccinatedControl,
  unvaccinatedControl,
  confint = c(0.025, 0.975)
)

odds_ratio_ve(
  vaccinatedCase,
  unvaccinatedCase,
  vaccinatedControl,
  unvaccinatedControl,
  confint = c(0.025, 0.975)
)

Arguments

vaccinatedCase

count of disease positive vaccine positive

unvaccinatedCase

count of disease positive vaccine negative

vaccinatedControl

count of disease negative vaccine positive

unvaccinatedControl

count of disease negative vaccine positive

confint

the confidence intervals

Value

a dataframe

a dataframe

Examples

tibble::tibble(
  N_vacc = 42240,
  N_unvacc = 42256,
  N_vacc_pn_pos = 49,
  N_unvacc_pn_pos = 90
) %>% dplyr::mutate(
  odds_ratio_ve(N_vacc_pn_pos, N_unvacc_pn_pos, N_vacc-N_vacc_pn_pos, N_unvacc-N_unvacc_pn_pos)
)
#> # A tibble: 1 × 10
#>   N_vacc N_unvacc N_vacc_pn_pos N_unvacc_pn_pos    OR OR.q.0.025 OR.q.0.975
#>    <dbl>    <dbl>         <dbl>           <dbl> <dbl>      <dbl>      <dbl>
#> 1  42240    42256            49              90 0.544      0.384      0.771
#> # ℹ 3 more variables: VE.OR <dbl>, VE.OR.q.0.025 <dbl>, VE.OR.q.0.975 <dbl>

tibble::tibble(
  N_vacc = 42240,
  N_unvacc = 42256,
  N_vacc_pn_pos = 49,
  N_unvacc_pn_pos = 90
) %>% dplyr::mutate(
  odds_ratio_ve(N_vacc_pn_pos, N_unvacc_pn_pos, N_vacc-N_vacc_pn_pos, N_unvacc-N_unvacc_pn_pos)
)
#> # A tibble: 1 × 10
#>   N_vacc N_unvacc N_vacc_pn_pos N_unvacc_pn_pos    OR OR.q.0.025 OR.q.0.975
#>    <dbl>    <dbl>         <dbl>           <dbl> <dbl>      <dbl>      <dbl>
#> 1  42240    42256            49              90 0.544      0.384      0.771
#> # ℹ 3 more variables: VE.OR <dbl>, VE.OR.q.0.025 <dbl>, VE.OR.q.0.975 <dbl>