Calculate an estimate of true prevalence for a single panel and components

Uses apparent prevalence, and uncertain estimates of test sensitivity and test specificity for the 3 methods described in Supplementary 2. This function works for a single panel per dataframe, multiple panels will need to call this function multiple times in a group_modify.

Usage

true_panel_prevalence(
  test_results = testerror:::.input_data,
  false_pos_controls = NULL,
  n_controls = NULL,
  false_neg_diseased = NULL,
  n_diseased = NULL,
  ...,
  sens = NULL,
  spec = NULL,
  panel_name = "Panel",
  confint = 0.95,
  method = c("rogan-gladen", "lang-reiczigel", "bayes"),
  na.rm = TRUE
)

Arguments

test_results

A dataframe containing the following columns:

id (character) - the patient identifier

result (logical) - the test result

test (factor) - the test type

Ungrouped

false_pos_controls

the number of positives that appeared in the specificity disease-free control group. These are by definition false positives. This is (1-specificity)*n_controls

n_controls

the number of controls in the specificity disease-free control group.

false_neg_diseased

the number of negatives that appeared in the sensitivity confirmed disease group. These are by definition false negatives. This is (1-sensitivity)*n_diseased

n_diseased

the number of confirmed disease cases in the sensitivity control group.

...

Arguments passed on to bayesian_panel_complex_model, bayesian_panel_simpler_model

panel_sens

the prior sensitivity of the panel as a beta_dist (optional)

panel_spec

the prior specificity of the panel as a beta_dist (optional)

fmt

a sprintf formatting string accepting 3 numbers

chains

A positive integer specifying the number of Markov chains. The default is 4.

iter

A positive integer specifying the number of iterations for each chain (including warmup). The default is 2000.

warmup

A positive integer specifying the number of warmup (aka burnin) iterations per chain. If step-size adaptation is on (which it is by default), this also controls the number of iterations for which adaptation is run (and hence these warmup samples should not be used for inference). The number of warmup iterations should be smaller than iter and the default is iter/2.

sens

the prior sensitivity of the test as a beta_dist.

spec

the prior specificity of the test as a beta_dist.

panel_name

the name of the panel for combined result

confint

confidence interval limits

method

one of:

"lang-reiczigel": Frequentist confidence limits

"bayes": Bayesian analysis

"rogan-gladen": Rogan gladen with uncertainty

na.rm

exclude patients with missing results

Value

A dataframe containing the following columns:

prevalence.label (character) - a fomatted label of the true prevalence estimate with CI

prevalence.lower (numeric) - the lower estimate

prevalence.median (numeric) - the median estimate

prevalence.method (character) - the method of estimation

prevalence.upper (numeric) - the upper estimate

test (character) - the name of the test or panel

Ungrouped

Examples

tmp = testerror:::panel_example()
#> Error in mutate(., n_components = n()): could not find function "mutate"
true_panel_prevalence(
  test_results = tmp$samples %>% dplyr::select(id,test,result = observed),
  false_pos_controls = tmp$performance$false_pos_controls,
  n_controls = tmp$performance$n_controls,
  false_neg_diseased = tmp$performance$false_neg_diseased,
  n_diseased = tmp$performance$n_diseased,
  method = "rogan-gladen"
)
#> Error in eval(expr, envir, enclos): object 'tmp' not found