Vectorised true prevalence estimates

Calculate an estimate of true prevalence from apparent prevalence, and uncertain estimates of test sensitivity and test specificity, using one of 3 methods.

Usage

true_prevalence(
  pos_obs,
  n_obs,
  false_pos_controls = NULL,
  n_controls = NULL,
  false_neg_diseased = NULL,
  n_diseased = NULL,
  confint = 0.95,
  method = c("lang-reiczigel", "rogan-gladen", "bayes"),
  ...,
  spec = NULL,
  sens = NULL
)

Arguments

pos_obs

the number of positive observations for a given test

n_obs

the number of observations for a given test

false_pos_controls

the number of positives that appeared in the specificity disease-free control group. These are by definition false positives. This is (1-specificity)*n_controls

n_controls

the number of controls in the specificity disease-free control group.

false_neg_diseased

the number of negatives that appeared in the sensitivity confirmed disease group. These are by definition false negatives. This is (1-sensitivity)*n_diseased

n_diseased

the number of confirmed disease cases in the sensitivity control group.

confint

confidence interval limits

method

one of:

"lang-reiczigel": Frequentist confidence limits: see prevalence_lang_reiczigel()

"rogan-gladen": Rogan gladen incorporating uncertainty with resampling: see uncertain_rogan_gladen()

"bayes": Bayesian analysis: see bayesian_component_simpler_model()

...

Arguments passed on to uncertain_rogan_gladen

samples

number fo random draws of sensitivity and specificity

fmt

a sprintf formatting string accepting 3 numbers

seed

set seed for reproducibility

spec

the prior specificity of the test as a beta_dist.

sens

the prior sensitivity of the test as a beta_dist.

Value

A dataframe containing the following columns:

prevalence.label (character) - a fomatted label of the true prevalence estimate with CI

prevalence.lower (numeric) - the lower estimate

prevalence.median (numeric) - the median estimate

prevalence.method (character) - the method of estimation

prevalence.upper (numeric) - the upper estimate

test (character) - the name of the test or panel

Ungrouped

Examples

true_prevalence(c(1:50), 200, 2, 800, 25, 75)
#> # A tibble: 50 × 15
#>    prevalence.lower prevalence.median prevalence.upper prevalence.method
#>               <dbl>             <dbl>            <dbl> <chr>            
#>  1          0                 0.00373           0.0420 lang-reiczigel   
#>  2          0                 0.0112            0.0534 lang-reiczigel   
#>  3          0                 0.0188            0.0644 lang-reiczigel   
#>  4          0.00313           0.0263            0.0751 lang-reiczigel   
#>  5          0.00787           0.0338            0.0856 lang-reiczigel   
#>  6          0.0128            0.0413            0.0959 lang-reiczigel   
#>  7          0.0179            0.0488            0.106  lang-reiczigel   
#>  8          0.0231            0.0564            0.116  lang-reiczigel   
#>  9          0.0284            0.0639            0.126  lang-reiczigel   
#> 10          0.0338            0.0714            0.136  lang-reiczigel   
#> # ℹ 40 more rows
#> # ℹ 11 more variables: prevalence.label <chr>, spec.median <dbl>,
#> #   spec.lower <dbl>, spec.upper <dbl>, spec.label <chr>, sens.median <dbl>,
#> #   sens.lower <dbl>, sens.upper <dbl>, sens.label <chr>, pos_obs <int>,
#> #   n_obs <dbl>
true_prevalence(c(1:10)*2, 200, 25, 800, 1, 6, method="rogan-gladen")
#> # A tibble: 10 × 15
#>    prevalence.median prevalence.lower prevalence.upper prevalence.label      
#>                <dbl>            <dbl>            <dbl> <chr>                 
#>  1            0               0                 0      0.00% [0.00% — 0.00%] 
#>  2            0               0                 0      0.00% [0.00% — 0.00%] 
#>  3            0               0                 0.0121 0.00% [0.00% — 1.21%] 
#>  4            0.0119          0                 0.0308 1.19% [0.00% — 3.08%] 
#>  5            0.0250          0.00864           0.0512 2.50% [0.86% — 5.12%] 
#>  6            0.0374          0.0195            0.0676 3.74% [1.95% — 6.76%] 
#>  7            0.0499          0.0311            0.0866 4.99% [3.11% — 8.66%] 
#>  8            0.0620          0.0423            0.112  6.20% [4.23% — 11.24%]
#>  9            0.0745          0.0547            0.124  7.45% [5.47% — 12.41%]
#> 10            0.0876          0.0673            0.159  8.76% [6.73% — 15.91%]
#> # ℹ 11 more variables: spec.median <dbl>, spec.lower <dbl>, spec.upper <dbl>,
#> #   spec.label <chr>, sens.median <dbl>, sens.lower <dbl>, sens.upper <dbl>,
#> #   sens.label <chr>, prevalence.method <chr>, pos_obs <dbl>, n_obs <dbl>
true_prevalence(c(1:10)*2, 200, 5, 800, 1, 6, method="bayes")
#> 
#> SAMPLING FOR MODEL 'component-simpler' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 2e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.2 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.158783 seconds (Warm-up)
#> Chain 1:                0.152411 seconds (Sampling)
#> Chain 1:                0.311194 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'component-simpler' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 1.5e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.15 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.153795 seconds (Warm-up)
#> Chain 2:                0.14783 seconds (Sampling)
#> Chain 2:                0.301625 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'component-simpler' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 1.5e-05 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.15 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 0.160532 seconds (Warm-up)
#> Chain 3:                0.150666 seconds (Sampling)
#> Chain 3:                0.311198 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'component-simpler' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 1.5e-05 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 0.15 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 0.146729 seconds (Warm-up)
#> Chain 4:                0.147935 seconds (Sampling)
#> Chain 4:                0.294664 seconds (Total)
#> Chain 4: 
#> # A tibble: 10 × 15
#>    prevalence.median prevalence.lower prevalence.upper prevalence.label       
#>                <dbl>            <dbl>            <dbl> <chr>                  
#>  1            0.0120         0.000653           0.0515 1.20% [0.07% — 5.15%]  
#>  2            0.0236         0.00239            0.0720 2.36% [0.24% — 7.20%]  
#>  3            0.0368         0.00787            0.105  3.68% [0.79% — 10.50%] 
#>  4            0.0508         0.0170             0.127  5.08% [1.70% — 12.67%] 
#>  5            0.0642         0.0261             0.156  6.42% [2.61% — 15.58%] 
#>  6            0.0794         0.0352             0.188  7.94% [3.52% — 18.79%] 
#>  7            0.0919         0.0447             0.210  9.19% [4.47% — 21.02%] 
#>  8            0.107          0.0549             0.244  10.70% [5.49% — 24.36%]
#>  9            0.120          0.0660             0.266  12.01% [6.60% — 26.63%]
#> 10            0.133          0.0750             0.285  13.30% [7.50% — 28.50%]
#> # ℹ 11 more variables: prevalence.method <chr>, sens.median <dbl>,
#> #   sens.lower <dbl>, sens.upper <dbl>, sens.label <chr>, spec.median <dbl>,
#> #   spec.lower <dbl>, spec.upper <dbl>, spec.label <chr>, pos_obs <dbl>,
#> #   n_obs <dbl>