Rogan-Gladen true prevalence for panel with resampling

Uses resampling to incorporate uncertainty of sensitivity and specificity into an estimate of true prevalence from a given value of apparent prevalence.

Usage

uncertain_panel_rogan_gladen(
  panel_pos_obs,
  panel_n_obs,
  pos_obs,
  n_obs,
  false_pos_controls = NA,
  n_controls = NA,
  false_neg_diseased = NA,
  n_diseased = NA,
  ...,
  spec = beta_dist(p = n_controls - false_pos_controls + 1, q = false_pos_controls + 1),
  sens = beta_dist(p = n_diseased - false_neg_diseased + 1, q = false_neg_diseased + 1),
  panel_sens = uncertain_panel_sens_estimator(pos_obs = pos_obs, n_obs = n_obs, sens =
    sens, spec = spec, samples = samples, fit_beta = FALSE),
  panel_spec = uncertain_panel_spec(spec = spec, samples = samples, fit_beta = FALSE),
  samples = 1000
)

Arguments

panel_pos_obs

the number of positive observations for a given panel of tests

panel_n_obs

the number of observations for each component test

pos_obs

a vector of the number of positive observations for each component of a panel test

n_obs

a vector of the number of observations for each component test

false_pos_controls

the number of positives that appeared in the component tests of a specificity disease-free control group. These are by definition false positives. This is (1-specificity)*n_controls (ignored if spec is given)

n_controls

the number of controls in the specificity disease-free control group (ignored if spec is given).

false_neg_diseased

the number of negatives that appeared in the component tests of a sensitivity confirmed disease group. These are by definition false negatives. This is (1-sensitivity)*n_diseased (ignored if sens is given)

n_diseased

the number of confirmed disease cases in the sensitivity control group (ignored if sens is given)

...

not used

spec

the specificity of the test as a beta_dist_list (optional)

sens

the sensitivity of the component tests of test as a beta_dist_list (optional)

panel_sens

the sensitivity of the component tests of test as a set of samples (optional)

panel_spec

the specificity of the test as a set of samples (optional)

samples

number of random draws of sensitivity and specificity (optional - default 1000)

Value

the expected value of apparent prevalence

Details

This is not vectorised