Calculate an estimate of true prevalence for a single panel and components

Uses apparent prevalence, and uncertain estimates of test sensitivity and test specificity for the 3 methods described in Supplementary 2. This function works for a single panel per dataframe, multiple panels will need to call this function multiple times in a group_modify.

Usage

true_panel_prevalence(
  test_results = testerror::.input_data,
  false_pos_controls = NULL,
  n_controls = NULL,
  false_neg_diseased = NULL,
  n_diseased = NULL,
  ...,
  sens = NULL,
  spec = NULL,
  panel_name = "Panel",
  confint = 0.95,
  method = c("rogan-gladen", "lang-reiczigel", "bayes"),
  na.rm = TRUE
)

Arguments

test_results

A dataframe containing the following columns:

id (character) - the patient identifier

result (logical) - the test result

test (factor) - the test type

Ungrouped

false_pos_controls

the number of positives that appeared in the specificity disease-free control group. These are by definition false positives. This is (1-specificity)*n_controls

n_controls

the number of controls in the specificity disease-free control group.

false_neg_diseased

the number of negatives that appeared in the sensitivity confirmed disease group. These are by definition false negatives. This is (1-sensitivity)*n_diseased

n_diseased

the number of confirmed disease cases in the sensitivity control group.

...

Arguments passed on to uncertain_panel_rogan_gladen, prevalence_panel_lang_reiczigel, bayesian_panel_complex_model, bayesian_panel_true_prevalence_model, bayesian_panel_simpler_model, bayesian_panel_logit_model

samples

number fo random draws of sensitivity and specificity

fmt

a sprintf formatting string accepting 3 numbers

panel_sens

the prior sensitivity of the panel as a beta_dist (optional)

panel_spec

the prior specificity of the panel as a beta_dist (optional)

cache_result

save the result of the sampling in memory for the current session

chains

A positive integer specifying the number of Markov chains. The default is 4.

iter

A positive integer specifying the number of iterations for each chain (including warmup). The default is 2000.

warmup

A positive integer specifying the number of warmup (aka burnin) iterations per chain. If step-size adaptation is on (which it is by default), this also controls the number of iterations for which adaptation is run (and hence these warmup samples should not be used for inference). The number of warmup iterations should be smaller than iter and the default is iter/2.

model_type

The bayesian model used - one of: logit, simpler, complex; default logit

sens

the prior sensitivity of the test as a beta_dist.

spec

the prior specificity of the test as a beta_dist.

panel_name

the name of the panel for combined result

confint

confidence interval limits

method

one of:

"lang-reiczigel": Frequentist confidence limits

"bayes": Bayesian analysis

"rogan-gladen": Rogan gladen with uncertainty

na.rm

exclude patients with missing results

Value

A dataframe containing the following columns:

prevalence.label (character) - a fomatted label of the true prevalence estimate with CI

prevalence.lower (numeric) - the lower estimate

prevalence.median (numeric) - the median estimate

prevalence.method (character) - the method of estimation

prevalence.upper (numeric) - the upper estimate

test (character) - the name of the test or panel

Ungrouped

Examples

tmp = testerror:::panel_example()
true_panel_prevalence(
  test_results = tmp$samples %>% dplyr::select(id,test,result = observed),
  false_pos_controls = tmp$performance$false_pos_controls,
  n_controls = tmp$performance$n_controls,
  false_neg_diseased = tmp$performance$false_neg_diseased,
  n_diseased = tmp$performance$n_diseased,
  method = "rogan-gladen"
)
#> Warning: qbeta(a, *) =: x0 with |pbeta(x0,*) - alpha| = 0.06217 is not accurate
#> Warning: qbeta(a, *) =: x0 with |pbeta(x0,*) - alpha| = 0.06217 is not accurate
#> # A tibble: 11 × 16
#>    prevalence.median prevalence.lower prevalence.upper prevalence.label     
#>                <dbl>            <dbl>            <dbl> <chr>                
#>  1           0.0102           0.00377          0.0144  1.02% [0.38% — 1.44%]
#>  2           0.00580          0                0.00942 0.58% [0.00% — 0.94%]
#>  3           0.0142           0.00621          0.0202  1.42% [0.62% — 2.02%]
#>  4           0.00524          0                0.00997 0.52% [0.00% — 1.00%]
#>  5           0.00799          0                0.0158  0.80% [0.00% — 1.58%]
#>  6           0.00957          0                0.0160  0.96% [0.00% — 1.60%]
#>  7           0                0                0.00532 0.00% [0.00% — 0.53%]
#>  8           0.0152           0.00530          0.0221  1.52% [0.53% — 2.21%]
#>  9           0.0184           0.0172           0.0211  1.84% [1.72% — 2.11%]
#> 10           0.00134          0.00110          0.00175 0.13% [0.11% — 0.17%]
#> 11           0.0794           0.0615           0.0976  7.94% [6.15% — 9.76%]
#> # ℹ 12 more variables: spec.median <dbl>, spec.lower <dbl>, spec.upper <dbl>,
#> #   spec.label <chr>, sens.median <dbl>, sens.lower <dbl>, sens.upper <dbl>,
#> #   sens.label <chr>, prevalence.method <chr>, test <chr>, pos_obs <int>,
#> #   n_obs <int>