Vectorised true prevalence estimates

Calculate an estimate of true prevalence from apparent prevalence, and uncertain estimates of test sensitivity and test specificity, using one of 3 methods.

Usage

true_prevalence(
  pos_obs,
  n_obs,
  false_pos_controls = NULL,
  n_controls = NULL,
  false_neg_diseased = NULL,
  n_diseased = NULL,
  confint = 0.95,
  method = c("lang-reiczigel", "rogan-gladen", "bayes"),
  ...,
  sens = NULL,
  spec = NULL
)

Arguments

pos_obs

the number of positive observations for a given test

n_obs

the number of observations for a given test

false_pos_controls

the number of positives that appeared in the specificity disease-free control group. These are by definition false positives. This is (1-specificity)*n_controls (ignored if spec given)

n_controls

the number of controls in the specificity disease-free control group.

false_neg_diseased

the number of negatives that appeared in the sensitivity confirmed disease group. These are by definition false negatives. This is (1-sensitivity)*n_diseased (ignored if sens given)

n_diseased

the number of confirmed disease cases in the sensitivity control group.

confint

confidence limit width

method

one of:

"lang-reiczigel": Frequentist confidence limits

"rogan-gladen": Rogan gladen incorporating uncertainty with resampling

"bayes": Bayesian analysis

...

passed onto methods

sens

the sensitivity of the test as a beta_dist (optional)

spec

the specificity of the test as a beta_dist (optional)

Value

A dataframe containing the following columns:

prevalence.label (character) - a fomatted label of the true prevalence estimate with CI

prevalence.lower (numeric) - the lower estimate

prevalence.median (numeric) - the median estimate

prevalence.method (character) - the method of estimation

prevalence.upper (numeric) - the upper estimate

test (character) - the name of the test or panel Grouped by:

Examples

true_prevalence(c(1:50), 200, 2, 800, 25, 75)
#> # A tibble: 50 × 15
#>    prevalence.lower prevalence.median prevalence.upper prevalence.method
#>               <dbl>             <dbl>            <dbl> <chr>            
#>  1          0                 0.00376           0.0421 lang-reiczigel   
#>  2          0                 0.0113            0.0536 lang-reiczigel   
#>  3          0                 0.0188            0.0646 lang-reiczigel   
#>  4          0.00317           0.0263            0.0753 lang-reiczigel   
#>  5          0.00792           0.0339            0.0858 lang-reiczigel   
#>  6          0.0129            0.0414            0.0961 lang-reiczigel   
#>  7          0.0179            0.0489            0.106  lang-reiczigel   
#>  8          0.0232            0.0565            0.116  lang-reiczigel   
#>  9          0.0285            0.0640            0.126  lang-reiczigel   
#> 10          0.0339            0.0715            0.136  lang-reiczigel   
#> # ℹ 40 more rows
#> # ℹ 11 more variables: prevalence.label <chr>, spec.median <dbl>,
#> #   spec.lower <dbl>, spec.upper <dbl>, spec.label <chr>, sens.median <dbl>,
#> #   sens.lower <dbl>, sens.upper <dbl>, sens.label <chr>, pos_obs <int>,
#> #   n_obs <dbl>
true_prevalence(c(1:10)*2, 200, 25, 800, 4.5, 5.5)
#> # A tibble: 10 × 15
#>    prevalence.lower prevalence.median prevalence.upper prevalence.method
#>               <dbl>             <dbl>            <dbl> <chr>            
#>  1                0            0                0.0559 lang-reiczigel   
#>  2                0            0                0.0959 lang-reiczigel   
#>  3                0            0                0.162  lang-reiczigel   
#>  4                0            0.0581           0.247  lang-reiczigel   
#>  5                0            0.125            0.342  lang-reiczigel   
#>  6                0            0.191            0.442  lang-reiczigel   
#>  7                0            0.257            0.544  lang-reiczigel   
#>  8                0            0.324            0.647  lang-reiczigel   
#>  9                0            0.390            0.751  lang-reiczigel   
#> 10                0            0.457            0.856  lang-reiczigel   
#> # ℹ 11 more variables: prevalence.label <chr>, spec.median <dbl>,
#> #   spec.lower <dbl>, spec.upper <dbl>, spec.label <chr>, sens.median <dbl>,
#> #   sens.lower <dbl>, sens.upper <dbl>, sens.label <chr>, pos_obs <dbl>,
#> #   n_obs <dbl>